fix: bcrypt on devcontainers + fix race condition on seed
