+++ app/seed/run.ts
@@ -1,0 +1,8 @@
+import { seed } from "./seed";
+
+seed()
+  .then(() => console.log("Database seeded"))
+  .catch((err) => {
+    console.error("An error happened while seeding the database");
+    console.error(err);
+  });
