+++ package.json
@@ -9,7 +9,7 @@
     "@heroicons/react": "^2.2.0",
     "@tailwindcss/forms": "^0.5.10",
     "autoprefixer": "10.4.20",
-    "bcrypt": "^5.1.1",
+    "bcryptjs": "^3.0.2",
     "clsx": "^2.1.1",
     "next": "15.4.0-canary.17",
     "next-auth": "5.0.0-beta.25",
